uc-crux-llvm: Split the test suite into more modules
This will make the code in Utils easier to reuse in new tests in new modules.
